+++ src/common/collections/tenants/fieldset.js
@@ -20,6 +20,19 @@ const _defaultFieldSet = function( conf ){
             schema: false,
             dt_title: pwixI18n.label( I18N, 'list.label_th' )
         },
+        // the tenant managers as an array, maybe empty
+        {
+            name: 'managers',
+            schema: false,
+            dt_title: pwixI18n.label( I18N, 'list.managers_th' ),
+            dt_render( data, type, rowData ){
+                let emails = [];
+                rowData.managers.forEach(( it ) => {
+                    emails.push( it.emails[0].address );
+                });
+                return emails.join( ', ' );
+            }
+        },
         // entity notes in tabular display
         {
             name: 'entity_notes',